fix: register to be landing page
